Refactor: split ingress code into a separate file
